requirements.txt
asgiref==3.2.7
Django==3.0.5
gunicorn==19.9.0
django-environ==0.4.5
django-tinymce==3.0.1
psycopg2==2.8.5
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.1
Unipath==1.1
django-extensions
pillow
